微服务拆分

+++ hmall/hm-service/src/main/java/com/hmall/service/impl/CartServiceImpl.java
@@ -113,14 +113,14 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private void checkCartsFull(Long userId) {
-        Long count = lambdaQuery().eq(Cart::getUserId, userId).count();
+        int count = lambdaQuery().eq(Cart::getUserId, userId).count();
         if (count >= 10) {
             throw new BizIllegalException(StrUtil.format("用户购物车课程不能超过{}", 10));
         }
     }
 
     private boolean checkItemExists(Long itemId, Long userId) {
-        Long count = lambdaQuery()
+        int count = lambdaQuery()
                 .eq(Cart::getUserId, userId)
                 .eq(Cart::getItemId, itemId)
                 .count();
